+++ cpp/src/EquivTrans.cpp
@@ -1,6 +1,6 @@
 #include "EquivTrans.hpp"
 
-EquivTrans::EquivTrans(const Transaction& trans, float weight) : myItems(trans), myWeight(weight)
+EquivTrans::EquivTrans(const Transaction& trans, int weight) : myItems(trans), myWeight(weight)
 {       
 }
